python/pyproject.toml
[tool.poetry]
name = "understory"
version = "0.0.92"
description = "Social web scaffolding."
authors = ["Angelo Gladding <angelo@ragt.ag>"]
license = "AGPL-3.0-or-later"
readme = "README.md"

[tool.poetry.urls]
Homepage = "https://understory.stream"

[tool.poetry.scripts]
loveliness = "understory.loveliness:main"
web = "understory.web.__main__:main"
build_understory = "main:build"

[tool.poetry.dependencies]
python = "^3.9"
argcomplete = "^1.12.3"
pendulum = "^2.1.2"
semver = "^2.13.0"
lxml = "^4.6.3"
Pygments = "^2.9.0"
emoji = "^1.2.0"
redis = "^3.5.3"
gunicorn = "^20.1.0"
hstspreload = "^2021.7.5"
requests = {extras = ["socks"], version = "^2.27.1"}
pyscreenshot = "^3.0"
Pillow = "^8.3.1"
PyVirtualDisplay = "^2.2"
selenium = "^4.1.0"
gevent = "^21.1.2"
jsonpatch = "^1.32"
pycryptodome = "^3.10.1"
Unidecode = "^1.2.0"
watchdog = "^2.1.3"
httpagentparser = "^1.9.1"
mimeparse = "^0.1.3"
acme-tiny = "^4.1.0"
vobject = "^0.9.6"
cssselect = "^1.1.0"
dnspython = "^2.1.0"
feedparser = "^6.0.8"
regex = "^2021.7.6"
rich = "^10.7.0"
microformats = "^0"
certifi = "^2021.10.8"
waitress = "^2.0.0"
vosk = "^0.3.32"
sounddevice = "^0.4.4"
gTTS = "^2.2.3"
pyttsx3 = "^2.90"
playsound = "^1.3.0"
PySide6 = { version = "^6.2.2", python = "<3.11" }
fifteen-ai = "^0.1.0"
coverage = "^6.3.1"

[tool.poetry.dev-dependencies]
ipython = "^8.0.1"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
webdriver-manager = "^3.4.2"
black = "^21.6b0"
pytest-gevent = "^1.1.0"
isort = "^5.9.2"
flake8 = "^3.9.2"
flake8-docstrings = "^1.6.0"
pep8-naming = "^0.12.0"
pydeps = "^1.10.12"
radon = "^5.1.0"
Nuitka = "^0.6.19"
coverage = "^6.3.1"
neovim = "^0.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
